feat(user-service): resolve pull request changes - modify kafka handler (seperated from fiber handler) - organize project structure Close#67
