+++ services/Dockerfile.auth
@@ -3,9 +3,13 @@ FROM golang:1.24 AS builder
 ARG SERVICE
 WORKDIR /bitka-exchange
 
-COPY ../../ . 
-RUN go mod download
+COPY ../go.work .
+COPY ./services/auth/go.mod ./services/auth/go.mod
+COPY ./services/account/go.mod ./services/account/go.mod
+COPY ./pkg/go.mod ./pkg/go.mod
 
+RUN go mod download
+COPY ./pkg ./pkg
 COPY ./services/${SERVICE} ./services/${SERVICE}
 
 RUN CGO_ENABLED=0 GOOS=linux go build \
@@ -18,4 +22,5 @@ ARG SERVICE
 ENV SERVICE=${SERVICE}
 
 COPY --from=builder /bin/${SERVICE}-service /bin/${SERVICE}-service
+
 CMD /bin/${SERVICE}-service
