+++ services/account/internal/delivery/event/server.go
@@ -1,0 +1,14 @@
+package event
+
+type Server struct {
+	consumer *Consumer
+}
+
+func NewServer(handler *Handler) *Server {
+	consumer := NewConsumer(handler)
+	return &Server{consumer: consumer}
+}
+
+func (s *Server) Start() {
+	s.consumer.Start()
+}
