+++ services/account/internal/delivery/http/dto/profile.go
@@ -1,0 +1,6 @@
+package dto
+
+type UpdateProfileRequest struct {
+	FullName  string `json:"full_name"`
+	AvatarURL string `json:"avatar_url"`
+}
