+++ services/auth/internal/delivery/http/dto/login.go
@@ -16,3 +16,5 @@ type LoginResponse struct {
 	RefreshToken string `json:"refresh_token"`
 	UserID       string `json:"user_id"`
 }
+
+
