+++ services/auth/internal/delivery/http/handler.go
@@ -37,19 +37,18 @@ func (h *AuthHandler) Register(c *fiber.Ctx) error {
 	if err := c.BodyParser(&req); err != nil {
 		return response.Error(c, fiber.StatusBadRequest, "Invalid request body")
 	}
-
 	if err := h.uc.Register(req.Email, req.Username, req.Password); err != nil {
 		return response.Error(c, fiber.StatusInternalServerError, "Registration failed")
 	}
 	if req.Email == "" {
-        return response.Error(c, fiber.StatusBadRequest, "Email is required")
-    }
-    if req.Username == "" {
-        return response.Error(c, fiber.StatusBadRequest, "Username is required")
-    }
-    if req.Password == "" {
-        return response.Error(c, fiber.StatusBadRequest, "Password is required")
-    }
+		return response.Error(c, fiber.StatusBadRequest, "Email is required")
+	}
+	if req.Username == "" {
+		return response.Error(c, fiber.StatusBadRequest, "Username is required")
+	}
+	if req.Password == "" {
+		return response.Error(c, fiber.StatusBadRequest, "Password is required")
+	}
 
 	return response.Success(c, "User registered successfully")
 }
